use int instead of long to calculate delta (like osmconvert does). This produces a different delta value for the 2nd lon. 	<node id="691940" lat="-39.2044208" lon="175.5375275" version="5"/> 	<node id="691943" lat="42.7857886" lon="-71.523316" version="2"/>
git-svn-id: http://svn.mkgmap.org.uk/splitter/branches/problem-list@238 838fabe5-b1a8-4b81-9896-7a84386980e9

+++ src/uk/me/parabola/splitter/O5mMapWriter.java
@@ -65,7 +65,7 @@ public class O5mMapWriter extends AbstractOSMWriter{
 	private long lastWayId;
 	private long lastRelId;
 	private long lastRef[];
-	private long lastLon,lastLat;
+	private int lastLon,lastLat;
 	
 	boolean isFirstNode = true;
 	boolean isFirstWay = true;
@@ -178,10 +178,10 @@ public class O5mMapWriter extends AbstractOSMWriter{
 		long delta = node.getId() - lastNodeId; lastNodeId = node.getId(); 
 		writeSignedNum(delta, stream);
 		stream.write(0x00); // no version info
-		long o5Lon = (long)(node.getLon() * FACTOR);
-		long o5Lat = (long)(node.getLat() * FACTOR);
-		long deltaLon = o5Lon - lastLon; lastLon = o5Lon;
-		long deltaLat = o5Lat - lastLat; lastLat = o5Lat;
+		int o5Lon = (int)(node.getLon() * FACTOR);
+		int o5Lat = (int)(node.getLat() * FACTOR);
+		int deltaLon = o5Lon - lastLon; lastLon = o5Lon;
+		int deltaLat = o5Lat - lastLat; lastLat = o5Lat;
 		writeSignedNum(deltaLon, stream);
 		writeSignedNum(deltaLat, stream);
 		writeTags(node, stream);
